Added benchmark for testing the NoThreadScheduler directly

+++ src/main/java/org/threadly/concurrent/benchmark/AbstractBenchmark.java
@@ -4,7 +4,7 @@ import java.security.SecureRandom;
 import java.util.Random;
 
 public abstract class AbstractBenchmark {
-  protected static final long RUN_TIME = 1000 * 60;
+  protected static final int RUN_TIME = 1000 * 60;
   
   protected static final Random RANDOM = new SecureRandom();
 }
